requirements.txt
gspread==6.2.1
google-auth==2.41.1
python-dotenv==1.1.1
